% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_friends.R
\name{densityplot}
\alias{densityplot}
\title{Densityplot of observed and imputed data}
\usage{
densityplot(...)
}
\arguments{
\item{...}{Any arguments passed to the function.}
}
\value{
The output of \link[mice:densityplot]{mice::densityplot} and a message about the \code{ggmice} equivalent.
}
\description{
Densityplot of observed and imputed data
}
\examples{
imp <- mice::mice(mice::nhanes, maxit = 1, printFlag = FALSE)
densityplot(imp)
}
