% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmice.R
\name{extract_mapping}
\alias{extract_mapping}
\title{Utils function to extract mapping variables}
\usage{
extract_mapping(data, vrbs, mapping_in)
}
\arguments{
\item{data}{Incomplete dataset or mids object.}

\item{vrbs}{Column names.}

\item{mapping_in}{Mapping provided to ggmice().}
}
\value{
Variable name from mapping_in argument matched on vrbs argument.
}
\description{
Utils function to extract mapping variables
}
\keyword{internal}
