% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmice-package.R
\docType{package}
\name{ggmice-package}
\alias{ggmice-package}
\title{ggmice: Visualizations for 'mice' with 'ggplot2'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Enhance a 'mice' imputation workflow with visualizations for incomplete and/or imputed data. The plotting functions produce 'ggplot' objects which may be easily manipulated or extended. Use 'ggmice' to inspect missing data, develop imputation models, evaluate algorithmic convergence, or compare observed versus imputed data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://amices.org/ggmice/}
  \item Report bugs at \url{https://github.com/amices/ggmice/issues}
}

}
\author{
\strong{Maintainer}: Hanne Oberman \email{h.i.oberman@uu.nl} (\href{https://orcid.org/0000-0003-3276-2141}{ORCID})

Authors:
\itemize{
  \item Hanne Oberman \email{h.i.oberman@uu.nl} (\href{https://orcid.org/0000-0003-3276-2141}{ORCID})
}

Other contributors:
\itemize{
  \item Utrecht University [funder]
  \item University Medical Centre Utrecht [funder]
  \item Thom Volker (\href{https://orcid.org/0000-0002-2408-7820}{ORCID}) [contributor]
  \item Gerko Vink (\href{https://orcid.org/0000-0001-9767-1924}{ORCID}) [contributor]
  \item Pepijn Vink (\href{https://orcid.org/0000-0001-6960-9904}{ORCID}) [contributor]
  \item Jamie Wallis (\href{https://orcid.org/0000-0003-2765-3813}{ORCID}) [contributor]
  \item Kyle Lang (\href{https://orcid.org/0000-0001-5340-7849}{ORCID}) [contributor]
}

}
\keyword{internal}
