% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmice.R
\name{ggmice}
\alias{ggmice}
\title{Plot incomplete or imputed data}
\usage{
ggmice(data = NULL, mapping = ggplot2::aes())
}
\arguments{
\item{data}{An incomplete dataset (of class \code{data.frame}), or an object of class \code{\link[mice:mids]{mice::mids}}.}

\item{mapping}{A list of aesthetic mappings created with \code{\link[ggplot2:aes]{ggplot2::aes()}}.}
}
\value{
An object of class \code{\link[ggplot2:ggplot]{ggplot2::ggplot}}. The \code{\link{ggmice}} function returns output
equivalent to \code{\link[ggplot2:ggplot]{ggplot2::ggplot}} output, with a few important exceptions:
\itemize{
\item The theme is set to \code{\link{theme_mice}}.
\item The color and fill scales are set to the \code{\link[mice:mdc]{mice::mdc}} colors (see below).
\item The \code{colour} aesthetic is set to \code{.where}, an internally defined variable which distinguishes
observed data from missing data or imputed data (for incomplete and imputed data, respectively).
\item The \code{fill} aesthetic is set to \code{.where}, an internally defined variable which distinguishes
observed data from missing data or imputed data (for incomplete and imputed data, respectively).
}
}
\description{
Plot incomplete or imputed data
}
\examples{
library(ggplot2)
# minimal example: scatterplot of incomplete and imputed data
dat <- mice::nhanes
ggmice(dat, aes(x = age, y = chl)) + geom_point()
imp <- mice::mice(dat, print = FALSE)
ggmice(imp, aes(x = age, y = chl)) + geom_point()

# more advanced functionality for incomplete data
# edit variable type for mixed incomplete data
dat$hyp <- factor(dat$hyp, levels = (1:2), labels = c("no hypertension", "hypertension"))
# scatterplot with categorical incomplete data
  ggmice(dat, aes(hyp, chl)) + geom_jitter(width = 0.1)
# incomplete data scatterplot faceted by categorical variable
ggmice(dat, aes(age, chl)) + geom_point() +
  facet_wrap(~ hyp, labeller = label_both)
# incomplete data scatterplot faceted by missing data indicator
ggmice(dat, aes(age, chl)) + geom_point() +
  facet_wrap(~ factor(is.na(hyp) == 0, labels = c("hyp observed", "hyp missing")))

# more advanced functionality for imputed data
# stripplot by imputation
ggmice(imp, aes(x = .imp, y = chl)) + geom_jitter(width = 0.25) +
  labs(x = "Imputation number")
# box plot by imputation
ggmice(imp, aes(x = .imp, y = chl)) + geom_boxplot() +
  labs(x = "Imputation number")
# density plot by imputation
ggmice(imp, aes(x = chl, group = .imp)) + geom_density()
# scatterplot faceted by imputation number
ggmice(imp, ggplot2::aes(x = age, y = bmi)) + ggplot2::geom_point() +
  facet_wrap(~ .imp)

}
\seealso{
See the \href{https://amices.org/ggmice/articles/ggmice.html}{\code{ggmice} vignette}
to use the \code{ggmice()} function on incomplete data or imputed data.
}
