% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern.R
\name{pat_to_chr}
\alias{pat_to_chr}
\title{Utils function to process missing data pattern}
\usage{
pat_to_chr(pat, ord = NULL)
}
\arguments{
\item{pat}{Numeric matrix with a missing data pattern.}

\item{ord}{Vector with variable names.}
}
\description{
Utils function to process missing data pattern
}
