% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_density.R
\name{plot_density}
\alias{plot_density}
\title{Plot the density of observed and imputed values}
\usage{
plot_density(data, vrb = "all", panels = FALSE)
}
\arguments{
\item{data}{An object of class \link[mice:mids-class]{mice::mids}.}

\item{vrb}{String, vector, or unquoted expression with variable name(s), default is "all".}

\item{panels}{Logical, vector of length 1 specifying whether the density plots should be broken into panels (TRUE) or not (FALSE)}
}
\value{
An object of class \link[ggplot2:ggplot]{ggplot2::ggplot}.
}
\description{
Plot the density of observed and imputed values
}
\examples{
imp <- mice::mice(mice::nhanes, print = FALSE)
plot_density(data = imp)
}
