% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_groups.R
\name{plot_groups}
\alias{plot_groups}
\title{(Clustered) missingness proportion per variable}
\usage{
plot_groups(
  data,
  vrb = "all",
  label = FALSE,
  square = TRUE,
  diagonal = FALSE,
  rotate = FALSE,
  useNA = "no"
)
}
\arguments{
\item{data}{A dataset of class \code{data.frame}, \code{tibble}, or \code{matrix}.}

\item{vrb}{String, vector, or unquoted expression with variable name(s), default is "all".}

\item{label}{Logical indicating whether correlation values should be displayed.}

\item{square}{Logical indicating whether the plot tiles should be squares.}

\item{diagonal}{Logical indicating whether the correlation of each variable with itself should be displayed.}

\item{rotate}{Logical indicating whether the variable name labels should be rotated 90 degrees.}

\item{useNA}{String, indicating whether to include a separate level for missing values. Options are "no" to exclude them from the plot, "both" to include both variables and missings and "only" to only include the 'NA' category on the y-axis.}
}
\value{
An object of class \link[ggplot2:ggplot]{ggplot2::ggplot}.
}
\description{
With this function, you can quickly check the number op subjects per level for
categorical variables and in total for continuous data. It creates a heat map
in which red means little subjects per group and blue numerous subjects per group.
}
\examples{
plot_groups(mice::boys, rotate = T, useNA = "no")
plot_groups(mice::boys, rotate = T, label = TRUE, useNA = "only")
plot_groups(mice::boys, rotate = T, useNA = "both")
}
