% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwplot.R
\name{plot_imps}
\alias{plot_imps}
\title{Title}
\usage{
plot_imps(imp, x, y = NULL)
}
\arguments{
\item{imp}{An object of class \code{mids}}

\item{x}{A variable name}

\item{y}{An optional second variable name}
}
\value{
An object of class \code{ggplot}
}
\description{
Title
}
\examples{
imp <- mice::mice(mice::nhanes, print = FALSE)
plot_imps(imp, x = "bmi")
}
