% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_miss.R
\name{plot_miss}
\alias{plot_miss}
\title{Plot missingness in a dataset}
\usage{
plot_miss(data, vrb = "all", border = FALSE, row.breaks = nrow(data))
}
\arguments{
\item{data}{An incomplete dataset of class \code{data.frame} or \code{matrix}.}

\item{vrb}{String, vector, or unquoted expression with variable name(s), default is "all".}

\item{border}{Logical indicating whether borders should be present between tiles.}

\item{row.breaks}{Optional numeric input specifying the number of breaks to be visualized on the y axis.}
}
\value{
An object of class \link[ggplot2:ggplot]{ggplot2::ggplot}.
}
\description{
Plot missingness in a dataset
}
\examples{
plot_miss(mice::nhanes)
}
