% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_moments.R
\name{plot_moments}
\alias{plot_moments}
\title{Plot the moments of the data and imputed values}
\usage{
plot_moments(data, vrb = "all")
}
\arguments{
\item{data}{An object of class \link[mice:mids-class]{mice::mids}.}

\item{vrb}{String, vector, or unquoted expression with variable name(s),
default is "all".}
}
\value{
An object of class \link[ggplot2:ggplot]{ggplot2::ggplot}.
}
\description{
Plot the moments of the data and imputed values
}
\details{
The \code{vrb} argument is "quoted" via \code{\link[rlang:defusing-advanced]{rlang::enexpr()}} and evaluated according
to \href{https://adv-r.hadley.nz/metaprogramming.html}{tidy evaluation principles}.
In practice, this technical nuance only affects users when passing an object
from the environment (e.g., a vector of variable names) to the \code{vrb} argument.
In such cases, the object must be "unquoted" via the \verb{!!} prefix operator.
}
\examples{
# create [mice::mids] object with [mice::mice()]
imp <- mice::mice(mice::nhanes, print = FALSE)

# plot moments for all imputed columns
plot_moments(imp)

# plot trace lines for specific columns by supplying a string or character vector
plot_moments(imp, "chl")
plot_moments(imp, c("chl", "hyp"))
# plot trace lines for specific columns by supplying unquoted variable names
plot_moments(imp, chl)
plot_moments(imp, c(chl, hyp))

# plot trace lines for specific columns by passing an object with variable names
# from the environment, unquoted with `!!`
my_variables <- c("chl", "hyp")
plot_moments(imp, !!my_variables)
# object with variable names must be unquoted with `!!`
try(plot_moments(imp, my_variables))

}
