% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pairwise.R
\name{plot_pairwise}
\alias{plot_pairwise}
\title{Plot pairwise variable-missingness indicator tests}
\usage{
plot_pairwise(data, vrb = "all", square = TRUE, rotate = FALSE)
}
\arguments{
\item{data}{A dataset of class \code{data.frame}, \code{tibble}, or \code{matrix}.}

\item{vrb}{String or vector with variable name(s), default is "all".}

\item{square}{Logical indicating whether the plot tiles should be squares.}

\item{rotate}{Logical indicating whether the variable name labels should be rotated 90 degrees.}
}
\value{
An object of class \link[ggplot2:ggplot]{ggplot2::ggplot}.
}
\description{
Performs t-tests and Fisher's exact tests to compare variables and their missing data indicators in a matrix format.
}
\examples{
plot_pairwise(mice::nhanes, rotate = TRUE)
}
