% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.pred.R
\name{plot_quickpred}
\alias{plot_quickpred}
\title{Title}
\usage{
plot_quickpred(dat, ...)
}
\arguments{
\item{dat}{An incomplete data.frame}

\item{...}{Optional additional arguments to \code{mice::quickpred}}
}
\value{
An object of class \code{ggplot}
}
\description{
Title
}
\examples{
plot_quickpred(mice::nhanes)
}
