% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{plot_test}
\alias{plot_test}
\title{Title}
\usage{
plot_test(dat)
}
\arguments{
\item{dat}{An incomplete data.frame}
}
\value{
An object of class \code{ggplot}
}
\description{
Title
}
\examples{
plot_corr(mice::nhanes)
}
