% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_friends.R
\name{stripplot}
\alias{stripplot}
\title{Stripplot of observed and imputed data}
\usage{
stripplot(...)
}
\arguments{
\item{...}{Any arguments passed to the function.}
}
\value{
The output of \link[mice:stripplot]{mice::stripplot} and a message about the \code{ggmice} equivalent.
}
\description{
Stripplot of observed and imputed data
}
\examples{
imp <- mice::mice(mice::nhanes, maxit = 1, printFlag = FALSE)
stripplot(imp)
}
