% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_mice}
\alias{theme_mice}
\title{Theme for \link{mice} style \link[ggplot2:ggplot]{ggplot2::ggplot} objects}
\usage{
theme_mice()
}
\value{
A \link{ggplot2} theme.
}
\description{
Theme for \link{mice} style \link[ggplot2:ggplot]{ggplot2::ggplot} objects
}
