% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_minimice}
\alias{theme_minimice}
\title{Minimal theme for \link{mice} style \link[ggplot2:ggplot]{ggplot2::ggplot} objects}
\usage{
theme_minimice()
}
\value{
A \link{ggplot2} theme.
}
\description{
Minimal theme for \link{mice} style \link[ggplot2:ggplot]{ggplot2::ggplot} objects
}
