% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_friends.R
\name{bwplot}
\alias{bwplot}
\title{Box-and-whisker plot of observed and imputed data}
\usage{
bwplot(...)
}
\arguments{
\item{...}{Any arguments passed to the function.}
}
\value{
The output of \code{mice::bwplot(...)} and a message about the \code{ggmice} equivalent.
}
\description{
Box-and-whisker plot of observed and imputed data
}
\examples{
imp <- mice::mice(mice::nhanes, maxit = 1, printFlag = FALSE)
bwplot(imp)
}
