% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityplot.R
\name{densityplot}
\alias{densityplot}
\title{Densityplot of observed and imputed data}
\usage{
densityplot(imp, vrb = "all")
}
\arguments{
\item{imp}{A \code{mids} object, typically created by \code{mice()} or \code{mice.mids()}}

\item{vrb}{String or vector with variable name(s), default is "all"}
}
\value{
A list with one or more \code{ggplot} objects
}
\description{
Densityplot of observed and imputed data
}
\examples{
imp <- mice::mice(mice::nhanes, printFlag = FALSE)
densityplot(imp, "bmi")
}
