% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmice.R
\name{ggmice}
\alias{ggmice}
\title{Title}
\usage{
ggmice(data = NULL, mapping = ggplot2::aes())
}
\arguments{
\item{data}{An incomplete dataset or an object of class \code{mids}}

\item{mapping}{A list of aesthetic mappings created with \code{ggplot2::aes()}}
}
\value{
A ggplot object of class \code{gg}
}
\description{
Title
}
\examples{
imp <- mice::mice(mice::nhanes, print = FALSE)
ggmice(imp, ggplot2::aes(x = age, y = bmi))
}
