% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mids.R
\name{plot_chains}
\alias{plot_chains}
\title{Title}
\usage{
plot_chains(imp)
}
\arguments{
\item{imp}{An object of class \code{mids}}
}
\value{
An object of class \code{ggplot}
}
\description{
Title
}
\examples{
imp <- mice::mice(mice::nhanes, print = FALSE)
plot_chains(imp)
}
