% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{plot_corr}
\alias{plot_corr}
\title{Title Plot correlations between (incomplete) variables}
\usage{
plot_corr(dat, vrb = "all", label = FALSE, square = TRUE, diagonal = FALSE)
}
\arguments{
\item{dat}{An incomplete data.frame.}

\item{vrb}{String or vector with variable name(s), default is "all".}

\item{label}{Logical indicating whether correlation values should be displayed.}

\item{square}{Logical indicating whether the plot tiles should be squares (setting the plot height equal to the plot width).}

\item{diagonal}{Logical indicating whether the correlation of each variable with itself should be displayed.}
}
\value{
An object of class \code{ggplot}
}
\description{
Title Plot correlations between (incomplete) variables
}
\examples{
plot_corr(mice::nhanes, label = TRUE)
}
