% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{plot_corr}
\alias{plot_corr}
\title{Title Plot correlations between (incomplete) variables}
\usage{
plot_corr(dat, label = FALSE)
}
\arguments{
\item{dat}{An incomplete data.frame}

\item{label}{Logical indicating whether correlation values should be displayed}
}
\value{
An object of class \code{ggplot}
}
\description{
Title Plot correlations between (incomplete) variables
}
\examples{
plot_corr(mice::nhanes, label = TRUE)
}
