% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux.R
\name{plot_flux}
\alias{plot_flux}
\title{Influx and outflux plot of multivariate missing data patterns}
\usage{
plot_flux(dat, label = TRUE, caption = TRUE)
}
\arguments{
\item{dat}{An incomplete dataset of class \code{data.frame}, \code{tibble}, or \code{matrix}.}

\item{label}{Logical indicating whether variable names should be displayed within the plot (the default) or with colors in the legend.}

\item{caption}{Logical indicating whether the figure caption should be displayed.}
}
\value{
An object of class \code{ggplot}.
}
\description{
Influx and outflux plot of multivariate missing data patterns
}
\examples{
plot_flux(mice::nhanes)
}
