% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux.R
\name{plot_flux}
\alias{plot_flux}
\title{Plot the influx and outflux of a multivariate missing data pattern}
\usage{
plot_flux(data, vrb = "all", label = TRUE, caption = TRUE)
}
\arguments{
\item{data}{An incomplete dataset of class \code{data.frame} or \code{matrix}.}

\item{vrb}{String or vector with variable name(s), default is "all".}

\item{label}{Logical indicating whether variable names should be displayed within the plot (the default) or with colors in the legend.}

\item{caption}{Logical indicating whether the figure caption should be displayed.}
}
\value{
An object of class \code{ggplot2::ggplot}.
}
\description{
Plot the influx and outflux of a multivariate missing data pattern
}
\examples{
plot_flux(mice::nhanes)
}
