% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_miss.R
\name{plot_miss}
\alias{plot_miss}
\title{Plot missingness in a dataset}
\usage{
plot_miss(data, vrb = "all", grid = FALSE, square = FALSE, ordered = FALSE)
}
\arguments{
\item{data}{An incomplete dataset of class \code{data.frame} or \code{matrix}.}

\item{vrb}{String, vector, or unquoted expression with variable name(s), default is "all".}

\item{grid}{Logical indicating whether borders should be present between tiles.}

\item{square}{Logical indicating whether the plot tiles should be squares, defaults to squares.}

\item{ordered}{Logical indicating whether rows should be ordered according to their pattern.}
}
\value{
An object of class \link[ggplot2:ggplot]{ggplot2::ggplot}.
}
\description{
Plot missingness in a dataset
}
\examples{
plot_miss(mice::nhanes)
}
