% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern.R
\name{plot_pattern}
\alias{plot_pattern}
\title{Plot the missing data pattern of an incomplete dataset}
\usage{
plot_pattern(dat)
}
\arguments{
\item{dat}{An incomplete data.frame.}
}
\value{
An object of class \code{ggplot}
}
\description{
Plot the missing data pattern of an incomplete dataset
}
\examples{
plot_pattern(mice::nhanes)
}
