% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern.R
\name{plot_pattern}
\alias{plot_pattern}
\title{Plot the missing data pattern of an incomplete dataset}
\usage{
plot_pattern(data, vrb = "all", square = FALSE, rotate = FALSE, cluster = NULL)
}
\arguments{
\item{data}{An incomplete dataset of class \code{data.frame} or \code{matrix}.}

\item{vrb}{String or vector with variable name(s), default is "all".}

\item{square}{Logical indicating whether the plot tiles should be squares.}

\item{rotate}{Logical indicating whether the variable name labels should be rotated 90 degrees.}

\item{cluster}{Optional character string specifying which variable should be used for clustering (e.g., for multilevel data).}
}
\value{
An object of class \code{ggplot2::ggplot}.
}
\description{
Plot the missing data pattern of an incomplete dataset
}
\examples{
plot_pattern(mice::nhanes)
}
