% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{plot_pred}
\alias{plot_pred}
\title{Plot predictorMatrix for \code{mice::mice()} argument}
\usage{
plot_pred(pred, label = FALSE, square = TRUE, rotate = FALSE)
}
\arguments{
\item{pred}{A predictor matrix for \code{mice}.}

\item{label}{Logical indicating whether predictor matrix values should be displayed.}

\item{square}{Logical indicating whether the plot tiles should be squares.}

\item{rotate}{Logical indicating whether the variable name labels should be rotated 90 degrees.}
}
\value{
An object of class \code{ggplot}.
}
\description{
Plot predictorMatrix for \code{mice::mice()} argument
}
\examples{
pred <- mice::make.predictorMatrix(mice::nhanes)
plot_pred(pred, label = TRUE)
}
