% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{plot_pred}
\alias{plot_pred}
\title{Title Plot predictorMatrix for \code{mice::mice()} argument}
\usage{
plot_pred(pred, label = FALSE)
}
\arguments{
\item{pred}{A predictor matrix for \code{mice}}

\item{label}{Logical indicating whether predictor matrix values should be displayed}
}
\value{
An object of class \code{ggplot}
}
\description{
Title Plot predictorMatrix for \code{mice::mice()} argument
}
\examples{
pred <- mice::make.predictorMatrix(mice::nhanes)
plot_pred(pred)
}
