% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.pred.R
\name{plot_pred}
\alias{plot_pred}
\title{Title}
\usage{
plot_pred(pred)
}
\arguments{
\item{pred}{A predictor matrix for \code{mice}}
}
\value{
An object of class \code{ggplot}
}
\description{
Title
}
\examples{
pred <- mice::make.predictorMatrix(mice::nhanes)
plot_pred(pred)
}
