% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace.R
\name{plot_trace}
\alias{plot_trace}
\title{Plot the trace lines of the MICE algorithm for convergence evaluation}
\usage{
plot_trace(imp, vrb = "all")
}
\arguments{
\item{imp}{An object of class \code{mids}.}

\item{vrb}{String or vector with variable name(s), default is "all".}
}
\value{
An object of class \code{ggplot}.
}
\description{
Plot the trace lines of the MICE algorithm for convergence evaluation
}
\examples{
imp <- mice::mice(mice::nhanes, print = FALSE)
plot_trace(imp)
}
