% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace.R
\name{plot_trace}
\alias{plot_trace}
\title{Plot the trace lines of the imputation algorithm}
\usage{
plot_trace(data, vrb = "all")
}
\arguments{
\item{data}{An object of class \code{mice::mids}.}

\item{vrb}{String or vector with variable name(s), default is "all".}
}
\value{
An object of class \code{ggplot2::ggplot}.
}
\description{
Plot the trace lines of the imputation algorithm
}
\examples{
imp <- mice::mice(mice::nhanes, print = FALSE)
plot_trace(imp)
}
