% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_variance.R
\name{plot_variance}
\alias{plot_variance}
\title{Plot as a heatmap the scaled between imputation variance for every cell}
\usage{
plot_variance(object)
}
\arguments{
\item{object}{A \code{mice} generated multiply imputed data set of class \code{mids}.}
}
\value{
An object of class \code{ggplot}.
}
\description{
This function plots the cell-level between imputation variance. The function
scales the variances column-wise and plots the data image as a heatmap.
Darker red cells indicate more variance, lighter cells indicate less variance.
white cells represent observed cells or unobserved cells with zero between
imputation variance.
}
\examples{
library(mice)
library(dplyr)
library(tidyr)
library(ggplot2)
imp <- mice(nhanes)
varianceplot(imp)
}
