% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_variance.R
\name{plot_variance}
\alias{plot_variance}
\title{Plot the scaled between imputation variance for every cell as a heatmap}
\usage{
plot_variance(data, vrb = "all", rotate = FALSE, grid = FALSE, square = FALSE)
}
\arguments{
\item{data}{A multiply imputed object of class \link[mice:mids]{mice::mids}.}

\item{vrb}{String, vector, or unquoted expression with variable name(s), default is "all".}

\item{rotate}{Logical indicating whether the variable name labels should be rotated 90 degrees.}

\item{grid}{Logical indicating whether borders should be present between tiles.}

\item{square}{Logical indicating whether the plot tiles should be squares, defaults to squares.}
}
\value{
An object of class \link[ggplot2:ggplot]{ggplot2::ggplot}.
}
\description{
This function plots the cell-level between imputation variance. The function
scales the variances column-wise, without centering cf. \code{base::scale(center = FALSE)}
and plots the data image as a heatmap. Darker red cells indicate more variance,
lighter cells indicate less variance. White cells represent observed cells or unobserved cells with zero between
imputation variance.
}
\examples{
# impute a dataset
imp <- mice::mice(mice::nhanes, print = FALSE)

# plot correlations for all columns
plot_variance(imp)

# plot correlations for specific columns by supplying a character vector
plot_variance(imp, c("chl", "hyp"))

# plot correlations for specific columns by supplying unquoted variable names
plot_variance(imp, c(chl, hyp))

# plot correlations for specific columns by passing an object with variable names
# from the environment, unquoted with `!!`
my_variables <- c("chl", "hyp")
plot_variance(imp, !!my_variables)
# object with variable names must be unquoted with `!!`
try(plot_variance(imp, my_variables))

}
