% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_friends.R
\name{stripplot}
\alias{stripplot}
\title{Stripplot of observed and imputed data}
\usage{
stripplot(imp, vrb = "all")
}
\arguments{
\item{imp}{A \code{mids} object, typically created by \code{mice::mice()} or \code{mice::mice.mids()} (see \code{?mice::mids}).}

\item{vrb}{String or vector with variable name(s), default is "all".}
}
\value{
A (list with) \code{ggplot} object(s).
}
\description{
Stripplot of observed and imputed data
}
\examples{
imp <- mice::mice(mice::nhanes, printFlag = FALSE)
stripplot(imp, "bmi")
}
