% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stripplot.R
\name{stripplot}
\alias{stripplot}
\title{Stripplot of observed and imputed data}
\usage{
stripplot(imp, vrb = "all")
}
\arguments{
\item{imp}{A \code{mids} object, typically created by \code{mice()} or \code{mice.mids()}.}

\item{vrb}{String or vector with variable name(s), default is "all".}
}
\value{
A list with one or more \code{ggplot} objects.
}
\description{
Stripplot of observed and imputed data
}
\examples{
imp <- mice::mice(mice::nhanes, printFlag = FALSE)
stripplot(imp, "bmi")
}
