% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_friends.R
\name{xyplot}
\alias{xyplot}
\title{Scatterplot of observed and imputed data}
\usage{
xyplot(imp, x, y)
}
\arguments{
\item{imp}{A \code{mids} object, typically created by \code{mice::mice()} or \code{mice::mice.mids()} (see \code{?mice::mids}).}

\item{x, y}{String with a variable name.}
}
\value{
A \code{ggplot} object
}
\description{
Scatterplot of observed and imputed data
}
\examples{
imp <- mice::mice(mice::nhanes, printFlag = FALSE)
xyplot(imp, "age", "bmi")
}
