% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_friends.R
\name{xyplot}
\alias{xyplot}
\title{Scatterplot of observed and imputed data}
\usage{
xyplot(...)
}
\arguments{
\item{...}{Any arguments passed to the function.}
}
\value{
The output of \code{mice::xyplot(...)} and a message about the \code{ggmice} equivalent.
}
\description{
Scatterplot of observed and imputed data
}
\examples{
imp <- mice::mice(mice::nhanes, maxit = 1, printFlag = FALSE)
xyplot(imp, bmi ~ age)
}
